chatbot-backend/requirements.txt
flask
flask-cors
spacy
python-docx
